wip for temp read through adc (mcp3002)
finally found a built in library for the adc converter (had to get a new ADC the mcp3002 which I got for free from the hive). So I will likely not need the mcp3002.py file but keeping just in case. Rn I am working to get this temp sensor working with the new library because it just reads zero for the moment.